example auto

+++ src/main/java/frc/robot/subsystem/arm/ArmControl.java
@@ -1,5 +1,5 @@
 package frc.robot.subsystem.arm;
 
 public class ArmControl {
-    public int angle;
+    public double angle;
 }
